src/backend/requirements.txt
fastapi
gunicorn
requests
pillow
transformers
uvicorn
opencv-python-headless

--extra-index-url https://download.pytorch.org/whl/cpu
torch